Testing different amounts of layers.
