+++ src/lib.rs
@@ -292,248 +292,3 @@ impl<N> BitSetLike<N> for BitSet<N>
         (&self).contains(i)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::{BitSet, BitSetAnd, BitSetNot, BitSetLike};
-
-    #[test]
-    fn insert() {
-        let mut c = BitSet::default();
-        for i in 0..1_000 {
-            assert!(!c.add(i));
-            assert!(c.add(i));
-        }
-
-        for i in 0..1_000 {
-            assert!(c.contains(i));
-        }
-    }
-
-    #[test]
-    fn insert_100k() {
-        let mut c = BitSet::default();
-        for i in 0..100_000 {
-            assert!(!c.add(i));
-            assert!(c.add(i));
-        }
-
-        for i in 0..100_000 {
-            assert!(c.contains(i));
-        }
-    }
-    #[test]
-    fn remove() {
-        let mut c = BitSet::default();
-        for i in 0..1_000 {
-            assert!(!c.add(i));
-        }
-
-        for i in 0..1_000 {
-            assert!(c.contains(i));
-            assert!(c.remove(i));
-            assert!(!c.contains(i));
-            assert!(!c.remove(i));
-        }
-    }
-
-    #[test]
-    fn iter() {
-        let mut c = BitSet::default();
-        for i in 0..100_000 {
-            c.add(i);
-        }
-
-        let mut count = 0;
-        for (idx, i) in c.iter().enumerate() {
-            count += 1;
-            assert_eq!(idx, i as usize);
-        }
-        assert_eq!(count, 100_000);
-    }
-
-    #[test]
-    fn iter_odd_even() {
-        let mut odd = BitSet::default();
-        let mut even = BitSet::default();
-        for i in 0..100_000 {
-            if i % 2 == 1 {
-                odd.add(i);
-            } else {
-                even.add(i);
-            }
-        }
-
-        assert_eq!((&odd).iter().count(), 50_000);
-        assert_eq!((&even).iter().count(), 50_000);
-        assert_eq!(BitSetAnd::new(&odd, &even).iter().count(), 0);
-    }
-
-    #[test]
-    fn iter_random_add() {
-        use rand::{Rng, weak_rng};
-        let mut set = BitSet::default();
-        let mut rng = weak_rng();
-        let limit = 1_048_576;
-        let mut added = 0;
-        for _ in 0..(limit / 10) {
-            let index = rng.gen_range(0, limit);
-            if !set.add(index) {
-                added += 1;
-            }
-        }
-        assert_eq!(set.iter().count(), added as usize);
-    }
-
-    #[test]
-    fn iter_clusters() {
-        let mut set = BitSet::default();
-        for x in 0..8 {
-            let x = (x * 3) << (::BITS * 2); // scale to the last slot
-            for y in 0..8 {
-                let y = (y * 3) << (::BITS);
-                for z in 0..8 {
-                    let z = z * 2;
-                    set.add(x + y + z);
-                }
-            }
-        }
-        assert_eq!(set.iter().count(), 8usize.pow(3));
-    }
-
-    #[test]
-    fn not() {
-        let mut c = BitSet::default();
-        for i in 0..10_000 {
-            if i % 2 == 1 {
-                c.add(i);
-            }
-        }
-        let d = BitSetNot::new(c);
-        for (idx, i) in d.iter().take(5_000).enumerate() {
-            assert_eq!(idx * 2, i as usize);
-        }
-    }
-}
-
-#[cfg(all(test, feature="parallel"))]
-mod test_parallel {
-    use super::{BitSet, BitSetAnd, BitSetLike};
-    use rayon::iter::ParallelIterator;
-
-    #[test]
-    fn par_iter_one() {
-        let step = 5000;
-        let tests = 1_048_576 / step;
-        for n in 0..tests {
-            let n = n * step;
-            let mut set = BitSet::default();
-            set.add(n);
-            assert_eq!(set.par_iter().count(), 1);
-        }
-        let mut set = BitSet::default();
-        set.add(1_048_576 - 1);
-        assert_eq!(set.par_iter().count(), 1);
-    }
-
-    #[test]
-    fn par_iter_random_add() {
-        use rand::{Rng, weak_rng};
-        use std::collections::HashSet;
-        use std::sync::{Arc, Mutex};
-        let mut set = BitSet::default();
-        let mut check_set = HashSet::new();
-        let mut rng = weak_rng();
-        let limit = 1_048_576;
-        for _ in 0..(limit / 10) {
-            let index = rng.gen_range(0, limit);
-            set.add(index);
-            check_set.insert(index);
-        }
-        let check_set = Arc::new(Mutex::new(check_set));
-        let missing_set = Arc::new(Mutex::new(HashSet::new()));
-        set.par_iter()
-            .for_each(|n| {
-                let check_set = check_set.clone();
-                let missing_set = missing_set.clone();
-                let mut check = check_set.lock().unwrap();
-                if !check.remove(&n) {
-                    let mut missing = missing_set.lock().unwrap();
-                    missing.insert(n);
-                }
-            });
-        let check_set = check_set.lock().unwrap();
-        let missing_set = missing_set.lock().unwrap();
-        if !check_set.is_empty() && !missing_set.is_empty() {
-            panic!("There were values that didn't get iterated: {:?}
-            There were values that got iterated, but that shouldn't be: {:?}", *check_set, *missing_set);
-        }
-        if !check_set.is_empty() {
-            panic!("There were values that didn't get iterated: {:?}", *check_set);
-        }
-        if !missing_set.is_empty() {
-            panic!("There were values that got iterated, but that shouldn't be: {:?}", *missing_set);
-        }
-    }
-
-    #[test]
-    fn par_iter_odd_even() {
-        let mut odd = BitSet::default();
-        let mut even = BitSet::default();
-        for i in 0..100_000 {
-            if i % 2 == 1 {
-                odd.add(i);
-            } else {
-                even.add(i);
-            }
-        }
-
-        assert_eq!((&odd).par_iter().count(), 50_000);
-        assert_eq!((&even).par_iter().count(), 50_000);
-        assert_eq!(BitSetAnd::new(&odd, &even).par_iter().count(), 0);
-    }
-
-    #[test]
-    fn par_iter_clusters() {
-        use std::collections::HashSet;
-        use std::sync::{Arc, Mutex};
-        let mut set = BitSet::default();
-        let mut check_set = HashSet::new();
-        for x in 0..8 {
-            let x = (x * 3) << (::BITS * 2); // scale to the last slot
-            for y in 0..8 {
-                let y = (y * 3) << (::BITS);
-                for z in 0..8 {
-                    let z = z * 2;
-                    let index = x + y + z;
-                    set.add(index);
-                    check_set.insert(index);
-                }
-            }
-        }
-        let check_set = Arc::new(Mutex::new(check_set));
-        let missing_set = Arc::new(Mutex::new(HashSet::new()));
-        set.par_iter()
-            .for_each(|n| {
-                let check_set = check_set.clone();
-                let missing_set = missing_set.clone();
-                let mut check = check_set.lock().unwrap();
-                if !check.remove(&n) {
-                    let mut missing = missing_set.lock().unwrap();
-                    missing.insert(n);
-                }
-            });
-        let check_set = check_set.lock().unwrap();
-        let missing_set = missing_set.lock().unwrap();
-        if !check_set.is_empty() && !missing_set.is_empty() {
-            panic!("There were values that didn't get iterated: {:?}
-            There were values that got iterated, but that shouldn't be: {:?}", *check_set, *missing_set);
-        }
-        if !check_set.is_empty() {
-            panic!("There were values that didn't get iterated: {:?}", *check_set);
-        }
-        if !missing_set.is_empty() {
-            panic!("There were values that got iterated, but that shouldn't be: {:?}", *missing_set);
-        }
-    }
-}
